4.7.4.5: Added check that selected os tz timezone exists also in ical database.
Timezone selection is based on operating system tz data but Orage uses internal
ical timezone data. So added test, which makes it possible to only show those
timezones that are available in ical. This test is not needed in the panel
plugin nor in globaltime as those use truely os timeinfo data.
Also fixed a big memory leak when switching display format.