4.8.1.13 Fixed Bug 7886 - .ics all-day-events span over an extra day
Full day events were shown one day too long. It was correct in the file
and in the appointment window, but all other places had it wrong.
Also fixed wrong alarm time for full day events. There still is a bug
if alarm is dependent on full day end time.